JSE_Bayesian_Regime_Swithcing/requirements.txt
# ============================================================
# JSE Regime Strategy — Dependencies
# ============================================================

# Data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
PyYAML>=6.0

# Statistical modelling
statsmodels>=0.14.0       # Classical Markov switching (Hamilton baseline)
pymc>=5.0.0               # Bayesian MCMC
pytensor>=2.18.0          # PyMC backend
scikit-learn>=1.4.0       # Preprocessing, clustering, metrics

# Backtesting
vectorbt>=0.26.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0            # Interactive charts for notebooks

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.66.0              # Progress bars
python-dotenv>=1.0.0      # Environment variables
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Code quality
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
